requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
pip==21.1.2
requests==2.28.1
setuptools==57.0.0
urllib3==1.26.13
wheel==0.36.2
python-dotenv==0.19.1
virtualenv==20.7.2